test: add authorization tests for check password feature
