+++ tests/test_unauthenticated_admin.py
@@ -1,12 +1,12 @@
 from snippets.lab4.example3_rpc_client import RemoteUserDatabase
-from snippets.lab4.users import User, Role, Token
+from snippets.lab4.users import User, Role, Token, Credentials
 from snippets.lab4.users.cryptography import DefaultSigner
 from snippets.lab4.example2_rpc_server import ServerStub, TEST_SECRET
 from datetime import datetime, timedelta
 import unittest
 import time
 
-SERVER_PORT = 9092
+SERVER_PORT = 9094
 
 ADMIN = User(
     username='luca',
@@ -24,7 +24,7 @@ USER = User(
     password='password'
 )
 
-class TestAuthenticatedAdminGetUser(unittest.TestCase):
+class TestUnauthenticatedAdminGetUser(unittest.TestCase):
     
     @classmethod
     def setUpClass(cls):
@@ -35,25 +35,28 @@ class TestAuthenticatedAdminGetUser(unittest.TestCase):
 
         cls.database.add_user(ADMIN)
         cls.database.add_user(USER)
-
-        token = cls.__create_valid_token_for_admin()
-        cls.database.set_token(token)
     
-    def test_authenticated_admin_gets_existing_user(self):
-        response = self.database.get_user(USER.username)
-        # two users are the same if they have exactly the same ids (username and email addresses)
-        self.assertEqual(USER, response)
-
-    def test_authenticated_admin_gets_non_existing_user(self):
-        fake_user = User(
-            username='mario',
-            emails={'mario@gmail.com'},
-            full_name='Mario Rossi',
-            role=Role.USER,
-            password='password'
-        )
+    def test_admin_without_token_get_user(self):
         with self.assertRaises(RuntimeError):
-            self.database.get_user(fake_user.username)
+            self.database.get_user(USER.username)
+    
+    def test_admin_without_token_check_password(self):
+        credentials = Credentials(ADMIN.username, ADMIN.password)
+        with self.assertRaises(RuntimeError):
+            self.database.check_password(credentials)
+
+    def test_admin_with_expired_token_get_user(self):
+        invalid_token = self.__create_invalid_token()
+        self.database.set_token(invalid_token)
+        with self.assertRaises(RuntimeError):
+            self.database.get_user(USER.username)
+
+    def test_admin_with_expired_token_check_password(self):
+        invalid_token = self.__create_invalid_token()
+        self.database.set_token(invalid_token)
+        credentials = Credentials(ADMIN.username, ADMIN.password)
+        with self.assertRaises(RuntimeError):
+            self.database.check_password(credentials)
 
     @classmethod
     def tearDownClass(cls):
@@ -61,8 +64,8 @@ class TestAuthenticatedAdminGetUser(unittest.TestCase):
         time.sleep(3)
         
     @classmethod
-    def __create_valid_token_for_admin(cls) -> Token:
+    def __create_invalid_token(cls) -> Token:
         signer = DefaultSigner(TEST_SECRET)
-        expiration = datetime.now() + timedelta(days=1)
+        expiration = datetime.now() - timedelta(days=1)
         signature = signer.sign(ADMIN, expiration)
         return Token(ADMIN, expiration, signature)
